Use the real key type size as initial agg state size
